pyproject.toml
[project]
name = "h2pcontrol-sdk"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.80.0",
    "beyer-labs-h2pcontrol-grpc-python>=1.80.0.1.20260406165239",
    "beyer-labs-h2pcontrol-protocolbuffers-pyi>=34.1.0.1.20260406165239",
    "beyer-labs-h2pcontrol-protocolbuffers-python>=34.1.0.1.20260406165239",
    "beyer-labs-h2pcontrol-community-nipunn1313-mypy-grpc>=4.0.0.1.20260407223931",
    "protobuf>=7.34.1",
    "pydantic-settings>=2.11.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "pyright>=1.1.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[[tool.uv.index]]
url = "https://buf.build/gen/python"
